+++ package.json
@@ -5,11 +5,18 @@
   "engines": {
     "node": "^4.0.0"
   },
-  "main": "src/index.js",
+  "main": "lib/index.js",
+  "jsnext:main": "src/index.js",
   "scripts": {
-    "test": "node tests/test.js"
+    "lint": "dignified-lint",
+    "test": "dignified-test node",
+    "build": "dignified-build",
+    "release": "dignified-release"
   },
-  "pre-commit": [],
+  "pre-commit": [
+    "lint",
+    "test"
+  ],
   "repository": {
     "type": "git",
     "url": "https://github.com/diasdavid/js-ipfs-kad-record-store.git"
@@ -25,6 +32,7 @@
   "homepage": "https://github.com/diasdavid/js-ipfs-kad-record-store",
   "devDependencies": {
     "abstract-record-store": "^0.1.1",
+    "dignified.js": "github:dignifiedquire/dignified.js",
     "libp2p-kad-routing": "^0.1.6",
     "libp2p-spdy": "^0.1.0",
     "libp2p-swarm": "^0.5.0",
@@ -34,8 +42,7 @@
     "peer-id": "^0.3.1",
     "peer-info": "^0.3.0",
     "pre-commit": "^1.1.1",
-    "standard": "^5.2.1",
-    "tape": "^4.2.0"
+    "tape": "^4.5.1"
   },
   "dependencies": {
     "ipld": "^0.4.0",
